stable version for Olivier, removed singletons and refactorised database reference for all API classes
