+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:pdg_app/model/message.dart';
@@ -14,7 +15,7 @@ void main() async {
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
-  IMessage msg = FirebaseMessage();
+  IMessage msg = FirebaseMessage(FirebaseFirestore.instance);
 
   final m1 = Message(id: '', fromId: 'alice', toId: 'bob', content: "HELLOW");
   msg.createMessage(m1);
